Remove old types
